pyproject.toml
[project]
name = "fastapi-mongo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.8",
    "motor>=3.7.0",
    "pydantic[dotenv,email]>=2.10.6",
    "pymongo>=4.11",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
